+++ src/FitnessTracker.Models/Fitness/GraphData/WorkoutGraphData.cs
@@ -2,7 +2,7 @@ namespace FitnessTracker.Models.Fitness.GraphData;
 
 public class WorkoutGraphData
 {
-    public double ExerciseMetaData { get; set; }
+    public decimal ExerciseMetaData { get; set; }
     public int XAxis { get; set; }
     public DateTimeOffset TimeOfExercise { get; set; }
 }
